Bump to v3 with new api for allowing forks
plus removed deprecated
plus zap fluffy `pkg/` dir
