+++ cmd/ovm/opened_versus_merged.go
@@ -11,11 +11,11 @@ import (
 	"time"
 
 	"github.com/google/go-github/v42/github"
-	"github.com/prince-chrismc/conan-center-index-pending-review/v2/internal"
-	"github.com/prince-chrismc/conan-center-index-pending-review/v2/internal/charts"
-	"github.com/prince-chrismc/conan-center-index-pending-review/v2/internal/duration"
-	"github.com/prince-chrismc/conan-center-index-pending-review/v2/internal/stats"
-	"github.com/prince-chrismc/conan-center-index-pending-review/v2/pkg/pending_review"
+	"github.com/prince-chrismc/conan-center-index-pending-review/v3/internal"
+	"github.com/prince-chrismc/conan-center-index-pending-review/v3/internal/charts"
+	"github.com/prince-chrismc/conan-center-index-pending-review/v3/internal/duration"
+	"github.com/prince-chrismc/conan-center-index-pending-review/v3/internal/stats"
+	"github.com/prince-chrismc/conan-center-index-pending-review/v3/pending_review"
 	"github.com/wcharczuk/go-chart/v2"
 )
