+++ internal/charts/line_chart.go
@@ -5,7 +5,7 @@ import (
 	"sort"
 	"time"
 
-	"github.com/prince-chrismc/conan-center-index-pending-review/v2/internal/stats"
+	"github.com/prince-chrismc/conan-center-index-pending-review/v3/internal/stats"
 	"github.com/wcharczuk/go-chart/v2"
 	"github.com/wcharczuk/go-chart/v2/drawing"
 )
