+++ internal/format/merge_ready.go
@@ -3,7 +3,7 @@ package format
 import (
 	"fmt"
 
-	"github.com/prince-chrismc/conan-center-index-pending-review/v2/pkg/pending_review"
+	"github.com/prince-chrismc/conan-center-index-pending-review/v3/pending_review"
 )
 
 // ReadyToMerge formats the pull request summaries into a markdown table for those considered 'ready to merge'
