+++ internal/format/under_way.go
@@ -3,7 +3,7 @@ package format
 import (
 	"fmt"
 
-	"github.com/prince-chrismc/conan-center-index-pending-review/v2/pkg/pending_review"
+	"github.com/prince-chrismc/conan-center-index-pending-review/v3/pending_review"
 )
 
 // UnderReview formats the pull request summaries into a markdown table for those **not** considered 'ready to merge'
